# doc/api/swagger.yaml
---
swagger: "2.0"
info:
  description: ""
  version: "3.1.2"
  title: "get songs"
host: "localhost:3000"
basePath: "/api"
schemes:
- "http"
paths:
  /songs:
    get:
      summary: "allow users see songs"
      description: "allow users see songs"
      operationId: "songsGET"
      parameters:
      - name: "limit"
        in: "query"
        required: false
        type: "integer"
        maximum: 20
      - name: "offset"
        in: "query"
        required: false
        type: "integer"
        default: 0
      - name: "author"
        in: "query"
        required: false
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "object"
            properties:
              songs:
                type: "array"
                items:
                  $ref: "#/definitions/Song"
        401:
          description: "Unauthorized"
        422:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/GenericErrorModel"
      x-swagger-router-controller: "Default"
    post:
      summary: "allow user add new song"
      description: "allow user add new song"
      operationId: "songsPOST"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Song"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/SongRequest"
        401:
          description: "Unauthorized"
        422:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/GenericErrorModel"
      security:
      - bearer: []
      x-swagger-router-controller: "Default"
  /songs/{slug}:
    put:
      summary: "allow user to change songs"
      description: "allow user to change songs"
      operationId: "songsSlugPUT"
      parameters:
      - name: "slug"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/SongRequest"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/SongRequest"
        401:
          description: "Unauthorized"
        422:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/GenericErrorModel"
      security:
      - bearer: []
      x-swagger-router-controller: "Default"
    delete:
      summary: "allow user to delete songs"
      description: "allow user to delete songs"
      operationId: "songsSlugDELETE"
      parameters:
      - name: "slug"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
        401:
          description: "Unauthorized"
        422:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/GenericErrorModel"
      security:
      - bearer: []
      x-swagger-router-controller: "Default"
securityDefinitions:
  bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  SongRequest:
    type: "object"
    required:
    - "song"
    properties:
      song:
        $ref: "#/definitions/Song"
    example:
      song:
        artist: "BTS"
        title: "Save me"
        lyrics: "AAAA [Verse 1] Nan sumswigo sipeo i bami silheo Ijen kkaego sipeo\
          \ kkumsogi silheo Nae ane gathyeoseo nan jugeoisseo Don’t wanna be lonely\
          \ Just wanna be yours Wae iri kkamkkamhan geonji Niga eopsneun i goseun\
          \ Wiheomhajanha manggajin nae moseup Guhaejwo nal nado nal jabeul su eopseo\
          \ Su eopseo\n[Pre-Chorus] Nae simjangsoril deureobwa Jemeosdaero neol bureujanha\
          \ I kkaman eodum sogeseo Neoneun ireohge biccnanikka\n[Chorus] Geu soneul\
          \ naemireojwo save me save me I need your love before I fall fall Geu soneul\
          \ naemireojwo save me save me I need your love before I fall fall Geu soneul\
          \ naemireojwo save me save me (Oh oh, save me me me) Geu soneul naemireojwo\
          \ save me save me (Oh oh, save me me me) Save me save me (Oh oh, save me\
          \ me me)"
  Song:
    type: "object"
    required:
    - "artist"
    - "lyrics"
    - "title"
    properties:
      title:
        type: "string"
        example: "Save me"
      artist:
        type: "string"
        example: "BTS"
      lyrics:
        type: "string"
        example: "AAAA [Verse 1] Nan sumswigo sipeo i bami silheo Ijen kkaego sipeo\
          \ kkumsogi silheo Nae ane gathyeoseo nan jugeoisseo Don’t wanna be lonely\
          \ Just wanna be yours Wae iri kkamkkamhan geonji Niga eopsneun i goseun\
          \ Wiheomhajanha manggajin nae moseup Guhaejwo nal nado nal jabeul su eopseo\
          \ Su eopseo\n[Pre-Chorus] Nae simjangsoril deureobwa Jemeosdaero neol bureujanha\
          \ I kkaman eodum sogeseo Neoneun ireohge biccnanikka\n[Chorus] Geu soneul\
          \ naemireojwo save me save me I need your love before I fall fall Geu soneul\
          \ naemireojwo save me save me I need your love before I fall fall Geu soneul\
          \ naemireojwo save me save me (Oh oh, save me me me) Geu soneul naemireojwo\
          \ save me save me (Oh oh, save me me me) Save me save me (Oh oh, save me\
          \ me me)"
    example:
      artist: "BTS"
      title: "Save me"
      lyrics: "AAAA [Verse 1] Nan sumswigo sipeo i bami silheo Ijen kkaego sipeo kkumsogi\
        \ silheo Nae ane gathyeoseo nan jugeoisseo Don’t wanna be lonely Just wanna\
        \ be yours Wae iri kkamkkamhan geonji Niga eopsneun i goseun Wiheomhajanha\
        \ manggajin nae moseup Guhaejwo nal nado nal jabeul su eopseo Su eopseo\n\
        [Pre-Chorus] Nae simjangsoril deureobwa Jemeosdaero neol bureujanha I kkaman\
        \ eodum sogeseo Neoneun ireohge biccnanikka\n[Chorus] Geu soneul naemireojwo\
        \ save me save me I need your love before I fall fall Geu soneul naemireojwo\
        \ save me save me I need your love before I fall fall Geu soneul naemireojwo\
        \ save me save me (Oh oh, save me me me) Geu soneul naemireojwo save me save\
        \ me (Oh oh, save me me me) Save me save me (Oh oh, save me me me)"
  GenericErrorModel:
    type: "object"
    required:
    - "errors"
    properties:
      errors:
        $ref: "#/definitions/GenericErrorModel_errors"
  inline_response_200:
    properties:
      songs:
        type: "array"
        items:
          $ref: "#/definitions/Song"
    example:
      songs:
      - artist: "BTS"
        title: "Save me"
        lyrics: "AAAA [Verse 1] Nan sumswigo sipeo i bami silheo Ijen kkaego sipeo\
          \ kkumsogi silheo Nae ane gathyeoseo nan jugeoisseo Don’t wanna be lonely\
          \ Just wanna be yours Wae iri kkamkkamhan geonji Niga eopsneun i goseun\
          \ Wiheomhajanha manggajin nae moseup Guhaejwo nal nado nal jabeul su eopseo\
          \ Su eopseo\n[Pre-Chorus] Nae simjangsoril deureobwa Jemeosdaero neol bureujanha\
          \ I kkaman eodum sogeseo Neoneun ireohge biccnanikka\n[Chorus] Geu soneul\
          \ naemireojwo save me save me I need your love before I fall fall Geu soneul\
          \ naemireojwo save me save me I need your love before I fall fall Geu soneul\
          \ naemireojwo save me save me (Oh oh, save me me me) Geu soneul naemireojwo\
          \ save me save me (Oh oh, save me me me) Save me save me (Oh oh, save me\
          \ me me)"
      - artist: "BTS"
        title: "Save me"
        lyrics: "AAAA [Verse 1] Nan sumswigo sipeo i bami silheo Ijen kkaego sipeo\
          \ kkumsogi silheo Nae ane gathyeoseo nan jugeoisseo Don’t wanna be lonely\
          \ Just wanna be yours Wae iri kkamkkamhan geonji Niga eopsneun i goseun\
          \ Wiheomhajanha manggajin nae moseup Guhaejwo nal nado nal jabeul su eopseo\
          \ Su eopseo\n[Pre-Chorus] Nae simjangsoril deureobwa Jemeosdaero neol bureujanha\
          \ I kkaman eodum sogeseo Neoneun ireohge biccnanikka\n[Chorus] Geu soneul\
          \ naemireojwo save me save me I need your love before I fall fall Geu soneul\
          \ naemireojwo save me save me I need your love before I fall fall Geu soneul\
          \ naemireojwo save me save me (Oh oh, save me me me) Geu soneul naemireojwo\
          \ save me save me (Oh oh, save me me me) Save me save me (Oh oh, save me\
          \ me me)"
  GenericErrorModel_errors:
    required:
    - "body"
    properties:
      body:
        type: "array"
        items:
          type: "string"